int、short、float、double的全序列化支持
git-svn-id: http://code.alibabatech.com/svn/fastjson/trunk/fastjson@1442 fb902949-7fc4-4485-a80f-4d6ff335c879

+++ src/main/java/com/alibaba/fastjson/serializer/SerializeConfig.java
@@ -96,8 +96,8 @@ public class SerializeConfig extends IdentityHashMap<Type, ObjectSerializer> {
 
         put(Boolean.class, BooleanSerializer.instance);
         put(Character.class, CharacterSerializer.instance);
-        put(Byte.class, IntegerSerializer.instance);
-        put(Short.class, IntegerSerializer.instance);
+        put(Byte.class, ByteSerializer.instance);
+        put(Short.class, ShortSerializer.instance);
         put(Integer.class, IntegerSerializer.instance);
         put(Long.class, LongSerializer.instance);
         put(Float.class, FloatSerializer.instance);
